svr: separate tx code from data code
